Rename mock folder to avoid conflicts
