+++ odw/test/integration_test/etl/curated/test_appeal_event_curated_process.py
@@ -2,7 +2,7 @@ import mock
 import pytest
 import pyspark.sql.types as T
 from pyspark.sql import functions as F
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from odw.core.etl.transformation.curated.appeal_event_curated_process import (
     AppealEventCuratedProcess,
 )
