+++ odw/test/integration_test/etl/curated/test_appeal_has_curated_mipins_process.py
@@ -6,7 +6,7 @@ from odw.test.util.assertion import (
     assert_etl_result_successful,
     assert_dataframes_equal,
 )
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from odw.core.etl.transformation.curated.appeal_has_curated_mipins_process import (
     AppealHasCuratedMipinsProcess,
 )
