+++ odw/test/integration_test/etl/curated/test_appeal_service_user_curated_mipins_process.py
@@ -1,6 +1,6 @@
 import mock
 import pytest
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from pyspark.sql import DataFrame
 from pyspark.sql.types import LongType, StringType, StructField, StructType
 from odw.core.etl.transformation.curated.appeal_service_user_curated_mipins_process import (
