+++ odw/test/integration_test/etl/curated/test_entraid_mipins_curation_process.py
@@ -1,5 +1,5 @@
 import pytest
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from odw.core.etl.transformation.curated.entraid_mipins_curation_process import (
     EntraIDMIPINSCurationProcess,
 )
