+++ odw/test/integration_test/etl/curated/test_nsip_invoice_curated_process.py
@@ -1,5 +1,5 @@
 import mock
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from odw.test.util.assertion import (
     assert_dataframes_equal,
     assert_etl_result_successful,
