+++ odw/test/integration_test/etl/curated/test_nsip_representation_curated_process.py
@@ -1,4 +1,4 @@
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from odw.core.etl.transformation.curated.nsip_representation_curated_process import (
     NsipRepresentationCuratedProcess,
 )
