+++ odw/test/integration_test/etl/curated/test_pins_inspector_curated_process.py
@@ -5,7 +5,7 @@ from odw.test.util.assertion import (
     assert_dataframes_equal,
     assert_etl_result_successful,
 )
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from odw.core.etl.transformation.curated.pins_inspector_curated_process import (
     PinsInspectorCuratedProcess,
 )
