+++ odw/test/integration_test/etl/harmonised/test_aie_document_harmonisation_process.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 import mock
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 import pyspark.sql.types as T
 from odw.core.etl.transformation.harmonised.aie_document_harmonisation_process import (
     AieDocumentHarmonisationProcess,
