+++ odw/test/integration_test/etl/harmonised/test_appeal_event_harmonisation_process.py
@@ -2,7 +2,7 @@ from datetime import datetime
 import mock
 import pytest
 import pyspark.sql.types as T
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from odw.core.etl.transformation.harmonised.appeal_event_harmonisation_process import (
     AppealEventHarmonisationProcess,
 )
