+++ odw/test/integration_test/etl/harmonised/test_appeal_has_harmonisation_process.py
@@ -7,7 +7,7 @@ from odw.test.util.assertion import (
     assert_etl_result_successful,
     assert_dataframes_equal,
 )
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from odw.core.etl.transformation.harmonised.appeal_has_harmonisation_process import (
     AppealHasHarmonisationProcess,
 )
