+++ odw/test/integration_test/etl/harmonised/test_entraid_harmonisation_process.py
@@ -2,7 +2,7 @@ import hashlib
 from datetime import datetime
 
 import mock
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 import pyspark.sql.types as T
 
 from odw.core.etl.transformation.harmonised.entraid_harmonisation_process import (
