+++ odw/test/integration_test/etl/harmonised/test_horizon_pins_inspector_harmonisation_process.py
@@ -6,7 +6,7 @@ from odw.test.util.assertion import (
     assert_dataframes_equal,
     assert_etl_result_successful,
 )
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from odw.core.etl.transformation.harmonised.horizon_pins_inspector_harmonisation_process import (
     HorizonPinsInspectorHarmonisationProcess,
 )
