+++ odw/test/integration_test/etl/harmonised/test_ref_appeal_attribute_matrix_harmonisation_process.py
@@ -1,5 +1,5 @@
 import pytest
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from odw.core.etl.transformation.harmonised.appeal_attribute_matrix_harmonisation_process import (
     AppealAttributeMatrixHarmonisationProcess,
 )
