+++ odw/test/integration_test/etl/harmonised/test_service_user_harmonisation_process.py
@@ -1,6 +1,6 @@
 import mock
 import pytest
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from pyspark.sql import DataFrame
 from pyspark.sql.types import LongType, StringType, StructField, StructType
 from odw.core.etl.transformation.harmonised.service_user_harmonisation_process import (
