+++ odw/test/integration_test/etl/standardised/test_appeal_has_standardisation_process.py
@@ -4,7 +4,7 @@ from odw.test.util.assertion import (
     assert_etl_result_successful,
     assert_dataframes_equal,
 )
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from odw.core.etl.transformation.standardised.appeal_has_standardisation_process import (
     AppealHasStandardisationProcess,
 )
