+++ odw/test/integration_test/etl/standardised/test_entraid_standardisation_process.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 import mock
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 import pyspark.sql.types as T
 
 from odw.core.etl.transformation.standardised.entraid_standardisation_process import (
