+++ odw/test/integration_test/etl/standardised/test_listed_building_standardisation_process.py
@@ -1,6 +1,6 @@
 import mock
 import pytest
-import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
+import odw.test.util.mock_util.import_mock_notebook_utils  # noqa: F401
 from pyspark.sql.types import StringType, StructField, StructType
 from odw.core.etl.transformation.standardised.listed_building_standardisation_process import (
     ListedBuildingStandardisationProcess,
