+++ odw/test/integration_test/etl/test_service_bus_standardisation_process.py
@@ -1,4 +1,4 @@
-from odw.test.util.mock.import_mock_notebook_utils import notebookutils  # noqa: F401
+from odw.test.util.mock_util.import_mock_notebook_utils import notebookutils  # noqa: F401
 from odw.core.etl.transformation.standardised.service_bus_standardisation_process import (
     ServiceBusStandardisationProcess,
 )
