+++ odw/test/unit_test/conftest.py
@@ -1,4 +1,4 @@
-from odw.test.util.mock import import_mock_notebook_utils  # noqa: F401 - must be imported before configure_session to mock notebookutils
+from odw.test.util.mock_util import import_mock_notebook_utils  # noqa: F401 - must be imported before configure_session to mock notebookutils
 from odw.test.util.conftest_util import (
     configure_session,
 )  # noqa: F401
