+++ odw/test/unit_test/util/test_logging_util.py
@@ -1,4 +1,4 @@
-from odw.test.util.mock.import_mock_notebook_utils import notebookutils
+from odw.test.util.mock_util.import_mock_notebook_utils import notebookutils
 from odw.core.util.logging_util import LoggingUtil
 from opentelemetry.sdk._logs import LoggerProvider
 from azure.monitor.opentelemetry.exporter import AzureMonitorLogExporter
